Separate clipboard related tests from InteractiveWindowTests.cs
